+++ zerver/tests/test_realm.py
@@ -1437,6 +1437,7 @@ class RealmAPITest(ZulipTestCase):
             realm_name_in_email_notifications_policy=UserProfile.REALM_NAME_IN_EMAIL_NOTIFICATIONS_POLICY_CHOICES,
             automatically_follow_topics_policy=UserProfile.AUTOMATICALLY_CHANGE_VISIBILITY_POLICY_CHOICES,
             automatically_unmute_topics_in_muted_streams_policy=UserProfile.AUTOMATICALLY_CHANGE_VISIBILITY_POLICY_CHOICES,
+            automatically_follow_topics_where_mentioned=[True, False],
         )
 
         vals = test_values.get(name)